chore: fix circular dependencies in Lookup Controller Code
