+++ packages/plugin-core/src/commands/NoteLookupCommand.ts
@@ -34,15 +34,17 @@ import {
 import {
   DendronQuickPickerV2,
   DendronQuickPickState,
-  LookupFilterType,
+  VaultSelectionMode,
+} from "../components/lookup/types";
+import {
   LookupNoteType,
   LookupNoteTypeEnum,
   LookupSelectionType,
   LookupSelectionTypeEnum,
   LookupSplitType,
   LookupSplitTypeEnum,
-  VaultSelectionMode,
-} from "../components/lookup/types";
+  LookupFilterType,
+} from "../components/lookup/ButtonTypes";
 import {
   node2Uri,
   OldNewLocation,
@@ -266,10 +268,7 @@ export class NoteLookupCommand
       lc.fuzzThreshold = copts.fuzzThreshold;
     }
 
-    if (!_.isUndefined(this.controller.view)) {
-      VSCodeUtils.setContext(DendronContext.SHOULD_SHOW_LOOKUP_VIEW, true);
-    }
-
+    VSCodeUtils.setContext(DendronContext.SHOULD_SHOW_LOOKUP_VIEW, true);
     VSCodeUtils.setContext(DendronContext.NOTE_LOOK_UP_ACTIVE, true);
 
     const { quickpick } = await lc.prepareQuickPick({
@@ -450,7 +449,9 @@ export class NoteLookupCommand
   cleanUp() {
     const ctx = "NoteLookupCommand:cleanup";
     Logger.debug({ ctx, msg: "enter" });
-    this.controller.onHide();
+    if (this._controller) {
+      this._controller.onHide();
+    }
     this.controller = undefined;
     HistoryService.instance().remove("lookup", "lookupProvider");
   }
@@ -490,7 +491,7 @@ export class NoteLookupCommand
     item: NoteQuickInput
   ): Promise<OnDidAcceptReturn | undefined> {
     const ctx = "acceptNewItem";
-    const picker = this.controller.quickpick;
+    const picker = this.controller.quickPick;
     const fname = this.getFNameForNewItem(item);
 
     const engine = ExtensionProvider.getEngine();
@@ -619,7 +620,7 @@ export class NoteLookupCommand
    */
   private getFNameForNewItem(item: NoteQuickInput) {
     if (this.isJournalButtonPressed()) {
-      return PickerUtilsV2.getValue(this.controller.quickpick);
+      return PickerUtilsV2.getValue(this.controller.quickPick);
     } else {
       return item.fname;
     }
@@ -680,11 +681,7 @@ export class NoteLookupCommand
   }
 
   private isJournalButtonPressed() {
-    const journalBtn = _.find(this.controller.state.buttons, (btn) => {
-      return btn.type === LookupNoteTypeEnum.journal;
-    });
-    const isPressed = journalBtn?.pressed;
-    return isPressed;
+    return this.controller.isJournalButtonPressed();
   }
 
   addAnalyticsPayload(opts?: CommandOpts, resp?: CommandOpts) {
