+++ packages/plugin-core/src/components/lookup/ButtonTypes.ts
@@ -1,14 +1,40 @@
 import _ from "lodash";
 import * as vscode from "vscode";
 import { QuickInputButton, ThemeIcon } from "vscode";
-import {
-  DendronQuickPickerV2,
-  LookupEffectType,
-  LookupFilterType,
-  LookupNoteType,
-  LookupSelectionType,
-  LookupSplitType,
-} from "./types";
+
+export type LookupFilterType = "directChildOnly";
+
+export enum LookupNoteTypeEnum {
+  "journal" = "journal",
+  "scratch" = "scratch",
+  "task" = "task",
+}
+export type LookupNoteType = LookupNoteTypeEnum;
+
+export enum LookupEffectTypeEnum {
+  "copyNoteLink" = "copyNoteLink",
+  "copyNoteRef" = "copyNoteRef",
+  "multiSelect" = "multiSelect",
+}
+export enum LookupSelectionTypeEnum {
+  "selection2link" = "selection2link",
+  "selectionExtract" = "selectionExtract",
+  "selection2Items" = "selection2Items",
+  "none" = "none",
+}
+export type LookupSelectionType =
+  | "selection2link"
+  | "selectionExtract"
+  | "selection2Items"
+  | "none";
+
+export enum LookupSplitTypeEnum {
+  "horizontal" = "horizontal",
+}
+export type LookupSplitType = "horizontal";
+
+export type LookupEffectType = "copyNoteLink" | "copyNoteRef" | "multiSelect";
+export type LookupNoteExistBehavior = "open" | "overwrite";
 
 export type ButtonType =
   | LookupEffectType
@@ -25,8 +51,6 @@ export type ButtonCategory =
   | "filter"
   | "effect"
   | "other";
-
-export type ButtonHandleOpts = { quickPick: DendronQuickPickerV2 };
 
 export function getButtonCategory(button: DendronBtn): ButtonCategory {
   if (isSelectionBtn(button)) {
@@ -79,7 +103,6 @@ export type IDendronQuickInputButton = QuickInputButton & {
   type: ButtonType;
   description: string;
   pressed: boolean;
-  onLookup: (payload: any) => Promise<void>;
 };
 
 type DendronBtnCons = {
@@ -96,10 +119,20 @@ export class DendronBtn implements IDendronQuickInputButton {
   public iconPathPressed: ThemeIcon;
   public type: ButtonType;
   public description: string;
-  public pressed: boolean;
+  private _pressed: boolean;
   public canToggle: boolean;
   public title: string;
   public opts: DendronBtnCons;
+
+  public get pressed() {
+    return this._pressed;
+  }
+
+  public set pressed(isPressed: boolean) {
+    if (this.canToggle) {
+      this._pressed = isPressed;
+    }
+  }
 
   onLookup = async (_payload: any) => {
     return;
@@ -111,7 +144,7 @@ export class DendronBtn implements IDendronQuickInputButton {
     this.iconPathPressed = new vscode.ThemeIcon(iconOn);
     this.type = type;
     this.description = description;
-    this.pressed = pressed || false;
+    this._pressed = pressed || false;
     this.title = title;
     this.canToggle = _.isUndefined(opts.canToggle) ? true : opts.canToggle;
     this.opts = opts;
@@ -120,15 +153,8 @@ export class DendronBtn implements IDendronQuickInputButton {
   clone(): DendronBtn {
     return new DendronBtn({
       ...this.opts,
+      pressed: this._pressed,
     });
-  }
-
-  async onEnable(_opts: ButtonHandleOpts): Promise<void> {
-    return undefined;
-  }
-
-  async onDisable(_opts: ButtonHandleOpts): Promise<void> {
-    return undefined;
   }
 
   get iconPath() {
