+++ packages/plugin-core/src/components/lookup/buttons.ts
@@ -1,32 +1,15 @@
-import {
-  ConfigUtils,
-  getSlugger,
-  MODIFIER_DESCRIPTIONS,
-  NoteProps,
-  NoteQuickInput,
-  NoteUtils,
-  TaskNoteUtils,
-} from "@dendronhq/common-all";
-import { WorkspaceUtils } from "@dendronhq/engine-server";
+import { MODIFIER_DESCRIPTIONS } from "@dendronhq/common-all";
 import _ from "lodash";
-import * as vscode from "vscode";
-import { DendronClientUtilsV2 } from "../../clientUtils";
-import { ExtensionProvider } from "../../ExtensionProvider";
-import { clipboard } from "../../utils";
-import { VSCodeUtils } from "../../vsCodeUtils";
-import { DendronBtn } from "./ButtonTypes";
-import { NotePickerUtils } from "./NotePickerUtils";
 import {
-  DendronQuickPickerV2,
+  DendronBtn,
   LookupEffectType,
-  LookupFilterType,
   LookupNoteType,
   LookupNoteTypeEnum,
   LookupSelectionType,
   LookupSplitType,
-  VaultSelectionMode,
-} from "./types";
-import { PickerUtilsV2 } from "./utils";
+  LookupFilterType,
+} from "./ButtonTypes";
+import { DendronQuickPickerV2 } from "./types";
 
 export type ButtonType =
   | LookupEffectType
@@ -93,76 +76,6 @@ function isSplitButton(button: DendronBtn) {
   return _.includes(["horizontal", "vertical"], button.type);
 }
 
-const selectionToNoteProps = async (opts: {
-  selectionType: string;
-  note: NoteProps;
-}) => {
-  const ext = ExtensionProvider.getExtension();
-  const resp = await VSCodeUtils.extractRangeFromActiveEditor();
-  const { document, range } = resp || {};
-  const { selectionType, note } = opts;
-  const { selection, text } = VSCodeUtils.getSelection();
-
-  switch (selectionType) {
-    case "selectionExtract": {
-      if (!_.isUndefined(document)) {
-        const ws = ExtensionProvider.getDWorkspace();
-        const lookupConfig = ConfigUtils.getCommands(ws.config).lookup;
-        const noteLookupConfig = lookupConfig.note;
-        const leaveTrace = noteLookupConfig.leaveTrace || false;
-        const body = "\n" + document.getText(range).trim();
-        note.body = body;
-        const { wsRoot, vaults } = ext.getDWorkspace();
-        // don't delete if original file is not in workspace
-        if (
-          !WorkspaceUtils.isPathInWorkspace({
-            wsRoot,
-            vaults,
-            fpath: document.uri.fsPath,
-          })
-        ) {
-          return note;
-        }
-        if (leaveTrace) {
-          const editor = VSCodeUtils.getActiveTextEditor();
-          const link = NoteUtils.createWikiLink({
-            note,
-            useVaultPrefix: DendronClientUtilsV2.shouldUseVaultPrefix(
-              ExtensionProvider.getEngine()
-            ),
-            alias: { mode: "title" },
-          });
-          await editor?.edit((builder) => {
-            if (!_.isUndefined(selection) && !selection.isEmpty) {
-              builder.replace(selection, `!${link}`);
-            }
-          });
-        } else {
-          await VSCodeUtils.deleteRange(document, range as vscode.Range);
-        }
-      }
-      return note;
-    }
-    case "selection2link": {
-      if (!_.isUndefined(document)) {
-        const editor = VSCodeUtils.getActiveTextEditor();
-        if (editor) {
-          await editor.edit((builder) => {
-            const link = note.fname;
-            if (!_.isUndefined(selection) && !selection.isEmpty) {
-              builder.replace(selection, `[[${text}|${link}]]`);
-            }
-          });
-        }
-      }
-      return note;
-    }
-    default: {
-      return note;
-    }
-  }
-};
-
 export class Selection2LinkBtn extends DendronBtn {
   static create(pressed?: boolean) {
     return new Selection2LinkBtn({
@@ -173,36 +86,6 @@ export class Selection2LinkBtn extends DendronBtn {
       type: "selection2link",
       pressed,
     });
-  }
-
-  async onEnable({ quickPick }: ButtonHandleOpts) {
-    quickPick.selectionProcessFunc = (note: NoteProps) => {
-      return selectionToNoteProps({
-        selectionType: "selection2link",
-        note,
-      });
-    };
-
-    quickPick.prevValue = quickPick.value;
-    const { text } = VSCodeUtils.getSelection();
-    const slugger = getSlugger();
-    quickPick.selectionModifierValue = slugger.slug(text!);
-    if (quickPick.noteModifierValue || quickPick.prefix) {
-      quickPick.value = NotePickerUtils.getPickerValue(quickPick);
-    } else {
-      quickPick.value = [
-        quickPick.rawValue,
-        NotePickerUtils.getPickerValue(quickPick),
-      ].join(".");
-    }
-    return;
-  }
-
-  async onDisable({ quickPick }: ButtonHandleOpts) {
-    quickPick.selectionProcessFunc = undefined;
-    quickPick.selectionModifierValue = undefined;
-    quickPick.value = NotePickerUtils.getPickerValue(quickPick);
-    return;
   }
 }
 
@@ -216,21 +99,6 @@ export class SelectionExtractBtn extends DendronBtn {
       type: "selectionExtract",
       pressed,
     });
-  }
-
-  async onEnable({ quickPick }: ButtonHandleOpts) {
-    quickPick.selectionProcessFunc = (note: NoteProps) => {
-      return selectionToNoteProps({
-        selectionType: "selectionExtract",
-        note,
-      });
-    };
-    return;
-  }
-
-  async onDisable({ quickPick }: ButtonHandleOpts) {
-    quickPick.selectionProcessFunc = undefined;
-    return;
   }
 }
 
@@ -250,21 +118,6 @@ export class Selection2ItemsBtn extends DendronBtn {
       canToggle,
     });
   }
-
-  async onEnable({ quickPick }: ButtonHandleOpts) {
-    const pickerItemsFromSelection =
-      NotePickerUtils.createItemsFromSelectedWikilinks();
-    quickPick.prevValue = quickPick.value;
-    quickPick.value = "";
-    quickPick.itemsFromSelection = pickerItemsFromSelection;
-    return;
-  }
-
-  async onDisable({ quickPick }: ButtonHandleOpts) {
-    quickPick.value = NotePickerUtils.getPickerValue(quickPick);
-    quickPick.itemsFromSelection = undefined;
-    return;
-  }
 }
 
 export class JournalBtn extends DendronBtn {
@@ -282,29 +135,6 @@ export class JournalBtn extends DendronBtn {
       pressed,
       canToggle,
     });
-  }
-
-  async onEnable({ quickPick }: ButtonHandleOpts) {
-    quickPick.modifyPickerValueFunc = () => {
-      return DendronClientUtilsV2.genNoteName("JOURNAL");
-    };
-    const { noteName, prefix } = quickPick.modifyPickerValueFunc();
-    quickPick.noteModifierValue = _.difference(
-      noteName.split("."),
-      prefix.split(".")
-    ).join(".");
-    quickPick.prevValue = quickPick.value;
-    quickPick.prefix = prefix;
-    quickPick.value = NotePickerUtils.getPickerValue(quickPick);
-    return;
-  }
-
-  async onDisable({ quickPick }: ButtonHandleOpts) {
-    quickPick.modifyPickerValueFunc = undefined;
-    quickPick.noteModifierValue = undefined;
-    quickPick.prevValue = quickPick.value;
-    quickPick.prefix = quickPick.rawValue;
-    quickPick.value = NotePickerUtils.getPickerValue(quickPick);
   }
 }
 
@@ -324,29 +154,6 @@ export class ScratchBtn extends DendronBtn {
       canToggle,
     });
   }
-
-  async onEnable({ quickPick }: ButtonHandleOpts) {
-    quickPick.modifyPickerValueFunc = () => {
-      return DendronClientUtilsV2.genNoteName("SCRATCH");
-    };
-    quickPick.prevValue = quickPick.value;
-    const { noteName, prefix } = quickPick.modifyPickerValueFunc();
-    quickPick.noteModifierValue = _.difference(
-      noteName.split("."),
-      prefix.split(".")
-    ).join(".");
-    quickPick.prefix = prefix;
-    quickPick.value = NotePickerUtils.getPickerValue(quickPick);
-    return;
-  }
-
-  async onDisable({ quickPick }: ButtonHandleOpts) {
-    quickPick.modifyPickerValueFunc = undefined;
-    quickPick.noteModifierValue = undefined;
-    quickPick.prevValue = quickPick.value;
-    quickPick.prefix = quickPick.rawValue;
-    quickPick.value = NotePickerUtils.getPickerValue(quickPick);
-  }
 }
 
 export class TaskBtn extends DendronBtn {
@@ -359,45 +166,6 @@ export class TaskBtn extends DendronBtn {
       type: LookupNoteTypeEnum.task,
       pressed,
     });
-  }
-
-  async onEnable({ quickPick }: ButtonHandleOpts) {
-    quickPick.modifyPickerValueFunc = () => {
-      return DendronClientUtilsV2.genNoteName(LookupNoteTypeEnum.task);
-    };
-    quickPick.prevValue = quickPick.value;
-    const { noteName, prefix } = quickPick.modifyPickerValueFunc();
-    quickPick.noteModifierValue = _.difference(
-      noteName.split("."),
-      prefix.split(".")
-    ).join(".");
-    quickPick.prefix = prefix;
-    quickPick.value = NotePickerUtils.getPickerValue(quickPick);
-    // If the lookup value ends up being identical to the current note, this will be confusing for the user because
-    // they won't be able to create a new note. This can happen with the default settings of Task notes.
-    // In that case, we add a trailing dot to suggest that they need to type something more.
-    const activeName = ExtensionProvider.getWSUtils().getActiveNote()?.fname;
-    if (quickPick.value === activeName) quickPick.value = `${quickPick.value}.`;
-    // Add default task note props to the created note
-    quickPick.onCreate = async (note) => {
-      note.custom = {
-        ...TaskNoteUtils.genDefaultTaskNoteProps(
-          note,
-          ConfigUtils.getTask(ExtensionProvider.getDWorkspace().config)
-        ).custom,
-        ...note.custom,
-      };
-      return note;
-    };
-    return;
-  }
-
-  async onDisable({ quickPick }: ButtonHandleOpts) {
-    quickPick.modifyPickerValueFunc = undefined;
-    quickPick.noteModifierValue = undefined;
-    quickPick.prevValue = quickPick.value;
-    quickPick.prefix = quickPick.rawValue;
-    quickPick.value = NotePickerUtils.getPickerValue(quickPick);
   }
 }
 
@@ -412,19 +180,6 @@ export class HorizontalSplitBtn extends DendronBtn {
       pressed,
     });
   }
-
-  async onEnable({ quickPick }: ButtonHandleOpts) {
-    quickPick.showNote = async (uri) =>
-      vscode.window.showTextDocument(uri, {
-        viewColumn: vscode.ViewColumn.Beside,
-      });
-    return;
-  }
-
-  async onDisable({ quickPick }: ButtonHandleOpts) {
-    quickPick.showNote = async (uri) => vscode.window.showTextDocument(uri);
-    return;
-  }
 }
 
 export class DirectChildFilterBtn extends DendronBtn {
@@ -437,18 +192,6 @@ export class DirectChildFilterBtn extends DendronBtn {
       type: "directChildOnly" as LookupFilterType,
       pressed,
     });
-  }
-
-  async onEnable({ quickPick }: ButtonHandleOpts) {
-    quickPick.showDirectChildrenOnly = true;
-    quickPick.filterMiddleware = (items: NoteQuickInput[]) => items;
-    return;
-  }
-
-  async onDisable({ quickPick }: ButtonHandleOpts) {
-    quickPick.showDirectChildrenOnly = false;
-    quickPick.filterMiddleware = undefined;
-    return;
   }
 }
 
@@ -468,15 +211,6 @@ export class MultiSelectBtn extends DendronBtn {
       canToggle,
     });
   }
-
-  async onEnable({ quickPick }: ButtonHandleOpts) {
-    quickPick.canSelectMany = this.pressed;
-    return;
-  }
-
-  async onDisable({ quickPick }: ButtonHandleOpts) {
-    quickPick.canSelectMany = this.pressed;
-  }
 }
 
 export class CopyNoteLinkBtn extends DendronBtn {
@@ -493,24 +227,6 @@ export class CopyNoteLinkBtn extends DendronBtn {
       // used to set it to FALSE.
       canToggle: true,
     });
-  }
-
-  async onEnable({ quickPick }: ButtonHandleOpts) {
-    quickPick.copyNoteLinkFunc = async (items: NoteProps[]) => {
-      const links = items.map((note) =>
-        NoteUtils.createWikiLink({ note, alias: { mode: "title" } })
-      );
-      if (_.isEmpty(links)) {
-        vscode.window.showInformationMessage(`no items selected`);
-      } else {
-        await clipboard.writeText(links.join("\n"));
-        vscode.window.showInformationMessage(`${links.length} links copied`);
-      }
-    };
-  }
-
-  async onDisable({ quickPick }: ButtonHandleOpts) {
-    quickPick.copyNoteLinkFunc = undefined;
   }
 }
 
@@ -529,39 +245,6 @@ export class VaultSelectButton extends DendronBtn {
 
   get tooltip(): string {
     return `${this.title}, status: ${this.pressed ? "always prompt" : "smart"}`;
-  }
-
-  setNextPicker({
-    quickPick,
-    mode,
-  }: {
-    quickPick: DendronQuickPickerV2;
-    mode: VaultSelectionMode;
-  }) {
-    quickPick.nextPicker = async (opts: { note: NoteProps }) => {
-      const { note } = opts;
-      const currentVault = PickerUtilsV2.getVaultForOpenEditor();
-      const vaultSelection = await PickerUtilsV2.getOrPromptVaultForNewNote({
-        vault: currentVault,
-        fname: note.fname,
-        vaultSelectionMode: mode,
-      });
-
-      if (_.isUndefined(vaultSelection)) {
-        vscode.window.showInformationMessage("Note creation cancelled.");
-        return;
-      }
-
-      return vaultSelection;
-    };
-  }
-
-  async onEnable({ quickPick }: ButtonHandleOpts) {
-    this.setNextPicker({ quickPick, mode: VaultSelectionMode.alwaysPrompt });
-  }
-
-  async onDisable({ quickPick }: ButtonHandleOpts) {
-    this.setNextPicker({ quickPick, mode: VaultSelectionMode.smart });
   }
 }
